requirements.txt
acres==0.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.5.0
asttokens==3.0.0
attrs==24.2.0
black==24.10.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
decorator==5.1.1
deepeval==1.1.6
distro==1.9.0
etelemetry==0.3.1
executing==2.1.0
faiss-cpu==1.9.0
filelock==3.16.1
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
httpx-sse==0.4.0
idna==3.10
importlib_resources==6.4.5
ipython==8.30.0
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.9
langchain-community==0.3.9
langchain-core==0.3.21
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langsmith==0.1.147
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.15.2
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==1.26.4
openai==1.56.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf>=3.20,<5.0
prov==2.0.1
pure_eval==0.2.3
puremagic==1.28
pyarrow==18.1.0
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydot==3.0.3
Pygments==2.18.0
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
rank-bm25==0.2.2
rdflib==6.3.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
scipy==1.14.1
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
streamlit==1.40.2
tenacity~=8.4.1
tiktoken==0.8.0
tokenize_rt==6.1.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
yarl==1.18.3
